move TRAINING_ITERATION
