+++ python/ray/tune/experiment/trial.py
@@ -20,7 +20,11 @@ from ray._private.dict import unflatten_dict
 from ray.air import CheckpointConfig
 from ray.air._internal.uri_utils import URI
 from ray.air._internal.checkpoint_manager import _TrackedCheckpoint, CheckpointStorage
-from ray.air.constants import EXPR_ERROR_PICKLE_FILE, EXPR_ERROR_FILE
+from ray.air.constants import (
+    EXPR_ERROR_PICKLE_FILE, 
+    EXPR_ERROR_FILE,
+    TRAINING_ITERATION,
+)
 
 import ray.cloudpickle as cloudpickle
 from ray.exceptions import RayActorError, RayTaskError
@@ -37,7 +41,6 @@ from ray.tune.result import (
     DONE,
     NODE_IP,
     PID,
-    TRAINING_ITERATION,
     TRIAL_ID,
     DEBUG_METRICS,
     TRIAL_INFO,
