+++ python/ray/tune/tests/execution/test_controller_search_alg_integration.py
@@ -6,11 +6,11 @@ import pytest
 import sys
 
 import ray
+from ray.air.constants import TRAINING_ITERATION
 from ray.air.execution import FixedResourceManager, PlacementGroupResourceManager
 from ray.tune import Experiment, PlacementGroupFactory
 from ray.tune.execution.tune_controller import TuneController
 from ray.tune.experiment import Trial
-from ray.tune.result import TRAINING_ITERATION
 from ray.tune.schedulers import FIFOScheduler, TrialScheduler
 from ray.tune.search import Searcher, ConcurrencyLimiter, Repeater, SearchGenerator
 from ray.tune.search._mock import _MockSuggestionAlgorithm
