+++ python/ray/tune/tests/test_function_api.py
@@ -6,6 +6,7 @@ import tempfile
 import unittest
 
 import ray
+from ray.air.constants import TRAINING_ITERATION
 from ray.rllib import _register_all
 
 from ray import tune
@@ -13,7 +14,7 @@ from ray.tune.logger import NoopLogger
 from ray.tune.execution.placement_groups import PlacementGroupFactory
 from ray.tune.trainable.util import TrainableUtil
 from ray.tune.trainable import with_parameters, wrap_function, FuncCheckpointUtil
-from ray.tune.result import DEFAULT_METRIC, TRAINING_ITERATION
+from ray.tune.result import DEFAULT_METRIC
 from ray.tune.schedulers import ResourceChangingScheduler
