+++ python/ray/tune/trainable/trainable.py
@@ -20,7 +20,11 @@ from ray.air.checkpoint import (
     Checkpoint,
     _DICT_CHECKPOINT_ADDITIONAL_FILE_KEY,
 )
-from ray.air.constants import TIMESTAMP, TIME_THIS_ITER_S
+from ray.air.constants import (
+    TIMESTAMP, 
+    TIME_THIS_ITER_S,
+    TRAINING_ITERATION,
+)
 from ray.tune.result import (
     DEBUG_METRICS,
     DEFAULT_RESULTS_DIR,
@@ -37,7 +41,6 @@ from ray.tune.result import (
     TIME_TOTAL_S,
     TIMESTEPS_THIS_ITER,
     TIMESTEPS_TOTAL,
-    TRAINING_ITERATION,
     TRIAL_ID,
     TRIAL_INFO,
 )
