+++ release/train_tests/horovod/train_horovod_multi_node_test.py
@@ -4,10 +4,10 @@ import time
 
 import ray
 from ray.air import ScalingConfig
+from ray.air.constants import TRAINING_ITERATION
 from ray.train.examples.horovod.horovod_example import (
     train_func as horovod_torch_train_func,
 )
-from ray.tune.result import TRAINING_ITERATION
 from ray.train.horovod.horovod_trainer import HorovodTrainer
 
 if __name__ == "__main__":
